requirements.txt
django>=5.1.7
llama-index-core>=0.10.0
llama-index-llms-groq>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
transformers>=4.38.0
torch>=2.2.0 